+++ src/helpers/index.ts
@@ -1,5 +1,9 @@
 export * from "./hex";
 export * from "./ipfs";
 export * from "./chainIdToChain";
-export { isChainSupported, doesChainSupportSplitManager } from "./supportedChains";
+export {
+  isChainSupported,
+  doesChainSupportSplitManager,
+  doesChainSupportMultiTokenFeeEscrow,
+} from "./supportedChains";
 export * from "./permissions";